llm_query_app/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
sentence-transformers
faiss-cpu
PyMuPDF
python-docx
requests
openai
